# Add extra inlining to speed up take