--- a/src/bitmap/mutable.rs
+++ b/src/bitmap/mutable.rs
@@ -348,6 +348,7 @@ impl MutableBitmap {
     }
 
     /// Creates a new [`MutableBitmap`] from an iterator of booleans.
+    #[inline]
     pub fn from_trusted_len_iter<I>(iterator: I) -> Self
     where
         I: TrustedLen<Item = bool>,
@@ -362,6 +363,7 @@ impl MutableBitmap {
     }
 
     /// Creates a new [`MutableBitmap`] from an iterator of booleans.
+    #[inline]
     pub fn try_from_trusted_len_iter<E, I>(iterator: I) -> std::result::Result<Self, E>
     where
         I: TrustedLen<Item = std::result::Result<bool, E>>,
@@ -372,6 +374,7 @@ impl MutableBitmap {
     /// Creates a new [`MutableBitmap`] from an falible iterator of booleans.
     /// # Safety
     /// The caller must guarantee that the iterator is `TrustedLen`.
+    #[inline]
     pub unsafe fn try_from_trusted_len_iter_unchecked<E, I>(
         mut iterator: I,
     ) -> std::result::Result<Self, E>