--- a/src/bitmap/mutable.rs
+++ b/src/bitmap/mutable.rs
@@ -348,7 +348,6 @@ impl MutableBitmap {
     }
 
     /// Creates a new [`MutableBitmap`] from an iterator of booleans.
-    #[inline]
     pub fn from_trusted_len_iter<I>(iterator: I) -> Self
     where
         I: TrustedLen<Item = bool>,
@@ -363,7 +362,6 @@ impl MutableBitmap {
     }
 
     /// Creates a new [`MutableBitmap`] from an iterator of booleans.
-    #[inline]
     pub fn try_from_trusted_len_iter<E, I>(iterator: I) -> std::result::Result<Self, E>
     where
         I: TrustedLen<Item = std::result::Result<bool, E>>,