--- a/src/buffer/mutable.rs
+++ b/src/buffer/mutable.rs
@@ -372,6 +372,7 @@ impl<T: NativeType> MutableBuffer<T> {
     /// # Safety
     /// This method assumes that the iterator's size is correct and is undefined behavior
     /// to use it on an iterator that reports an incorrect length.
+    #[inline]
     pub unsafe fn extend_from_trusted_len_iter_unchecked<I: Iterator<Item = T>>(
         &mut self,
         iterator: I,
@@ -461,6 +462,7 @@ impl<T: NativeType> MutableBuffer<T> {
     /// # Safety
     /// This method assumes that the iterator's size is correct and is undefined behavior
     /// to use it on an iterator that reports an incorrect length.
+    #[inline]
     pub unsafe fn try_from_trusted_len_iter_unchecked<
         E,
         I: Iterator<Item = std::result::Result<T, E>>,