@jorgecarleitao 2021-07-25
Uf, amazing. Thanks a lot, @Dandandan .  cc @ritchie46 , since this is relevant to Polars.